debug for dtype

+++ starVLA/model/framework/VLM4A/CosmosGR00T.py
@@ -187,5 +187,5 @@ class Cosmos_GR00T(baseframework):
         with torch.autocast("cuda", dtype=torch.float32):
             pred_actions = self.action_model.predict_action(last_hidden, state)
 
-        normalized_actions = pred_actions.detach().cpu().numpy()
+        normalized_actions = pred_actions.detach().float().cpu().numpy()
         return {"normalized_actions": normalized_actions}
