+++ starVLA/model/framework/VLM4A/QwenDual.py
@@ -220,7 +220,7 @@ class Qwen_Dual(baseframework):
         # Step 4: Action Expert Forward
         with torch.autocast("cuda", dtype=torch.float32):
             pred_actions = self.action_model.predict_action(last_hidden, state)  # (B, chunk_len, action_dim)
-        normalized_actions = pred_actions.detach().cpu().numpy()
+        normalized_actions = pred_actions.detach().float().cpu().numpy()
 
         return {"normalized_actions": normalized_actions}
